myproject/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
reportlab==4.0.7
PyQt5==5.15.10
matplotlib==3.8.2
requests==2.31.0
